requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sentence-transformers
scikit-learn
Pillow
numpy
